GCC static_cast bug fix

+++ source/engine/graphics/api/vulkan/RenderGraph.cpp
@@ -45,7 +45,7 @@ void SetViewportAndScissorAspectRatio(vk::CommandBuffer* cmd, const nc::Vector2&
 {
     const auto viewport = vk::Viewport{(dimensions.x - screenExtent.x) / 2, (dimensions.y - screenExtent.y) / 2, screenExtent.x, screenExtent.y, 0.0f, 1.0f};
     const auto extent = vk::Extent2D{static_cast<uint32_t>(screenExtent.x), static_cast<uint32_t>(screenExtent.y)};
-    const auto scissor = vk::Rect2D{vk::Offset2D{static_cast<int32_t>(dimensions.x - extent.width) / 2, static_cast<int32_t>(dimensions.y - extent.height) / 2}, extent};
+    const auto scissor = vk::Rect2D{vk::Offset2D{static_cast<int32_t>(dimensions.x - static_cast<float>(extent.width)) / 2, static_cast<int32_t>(dimensions.y - static_cast<float>(extent.height)) / 2}, extent};
     cmd->setViewport(0, 1, &viewport);
     cmd->setScissor(0, 1, &scissor);
 }
